backend/requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pandas>=2.1.0
python-dateutil>=2.8.2
aiohttp>=3.9.0
websockets>=12.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.4
schedule>=1.2.0
pydantic-settings>=2.1.0
